Initial commit
Ptolemy aims at making a web-based counterpart of the useful software "Kleopatra",
to ease the use of OpenPGP and make encryption more available to normal users.

At backend this project will use OpenPGP.js. Keyring will be stored in browser IndexedDB,
or when used as a Web extension, maybe locally.