Use OpenPGP.js to build the encrypted keyring
We simply use an OpenPGP private key as master key, and encrypt all localStorage items
with it. Just don't reinvent the wheels.

If feeling insecure, use scrypt to set the private key password in the future. Salt may be the random user id set in UserID of that private key.